requirements/common.txt
capmonster-python==2.2
flask==2.0.3
flask-pydantic==0.9.0
gunicorn==20.1.0
pydantic==1.9.0
pyotp==2.6.0
python-dotenv==0.19.2
retry==0.9.2
selenium==4.1.0
selenium-stealth==1.0.6
selenium-wire==4.6.0